pyproject.toml
[project]
name = "stripe-mock-tutorial"
version = "0.1.0"
description = ""
authors = [
    {name = "guilhermegouw",email = "guilherme.gouw@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "stripe (>=12.3.0,<13.0.0)",
    "python-decouple (>=3.8,<4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"
pytest-mock = "^3.14.1"
black = "^25.1.0"
flake8 = "^7.3.0"
isort = "^6.0.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--reuse-db --nomigrations"
